Fix schemagen bool encoding for false values
Extract schema generation logic into testable functions and fix a bug where
boolean fields with false values were not being encoded due to Empty() check
treating false as a zero value.

Changes:
- Extract generation logic from main.go into generator.go with testable functions
- Add comprehensive unit tests for bool encoding behavior
- Fix bool encoding to always encode values without Empty() check
- Maintain existing behavior for other field types (strings still use Empty())
